[typing] fixes issue with relative order of defined vs declared local vars
If a defined local is referenced to in the type of a declared one (e.g. in
a `sizeof` expression as part of an array length), the `Local_init` will put
its actual definition after the declaration (which occurs at the beginning of
the block, resulting in at least pretty-printing unparsable code. We generate
undeclared temp variables of the same type to overcome the issue, like in
bb30d0e4f8476