[Eva] Builtins: new function prepare_builtins called at the start of the analysis.
This function links the kernel functions to their builtins for a given analysis.
Checks that the builtin replacing a function has a compatible type and a
specification. These requirements are not checked each time a builtin is used.

The type of a builtin is now a function "unit -> typ" applied when the analysis
starts, allowing builtins to use types defined from the machdep, such as size_t.