[Eva] Fixes a precision bug in the subdivisions when a domain uses the oracle.
Resets the reference to the comparison function at the end of [do_subdiv].

The evaluation of an expression performs successively two subdivision runs: one
to improve the lower bound of the result and one to improve its higher bound.
A different order on the subdivision disjuncts is used for each run, but should
remain the same during one run. The order guides the subdivision to split the
disjunct for which the given bound is the least precise.

During a subdivision, if a domain uses the oracle on a non-linear expression,
its evaluation is also subdivided, and the reference is changed. When returning
to the initial subdivision, the order used between the disjuncts of the
subdivision is wrong and can lead to imprecise results.

This explains the precision improvement on tests with the octagon domain.