[Eva] Subdivision: slightly changes the strategy to cut a float interval in half.
If an interval contains positive and negative floating-point values, cuts the
interval at 0. Otherwise, uses the same strategy as before: cuts the interval
at the mathematical average of its bounds.

This strategy always reaches the maximal precision in one split on e*e squares,
which is the most common pattern for subdivisions.
In the tests, it decreases the precision on computations of e-e, where a split
at 0 is useless. However, this is not a common pattern in real codes.