[eacsl:codegen] Update inject_in_block to support the new clean and delete globals placements
Since `memory_clean` and `delete_globals` are added after the calls to
`inject_in_block`, this function must be updated to add the free locals
statements as last statements of the innermost block instead of before
the `memory_clean` statement in the innermost block. Additionnaly, this
fix harmonize the treatment between a classic function and the function
`main()`, and fix the issue frama-c/e-acsl#105.