[Eva] Improves the automatic loop unroll when temporary variables are introduced.
If the variable used in the exit condition of the loop is not incremented within
the loop, but is assigned to the value of another lvalue, the heuristic uses
this second lvalue instead.
This is especially useful to deal with some temporary variables introduced
by Frama-C to translate condition such as (i++ < l).