[lexer] tentative logic lexer hack for introducing generalized check
Just using `check` as a prefix to various keywords would lead to too many
ambiguities in the parser (notably in contracts, where you can have an empty
requires list followed by `check ensures ...` or a `check requires ...`. With
a lookahead of 1, it's difficult for the parser see which rule should apply.
Hence, the lexer will perform itself the lookahead and emit a proper token
(`CHECK_ENSURES` or `CHECK_REQUIRES` respectively).