[aorai] fix behavior generation for absence of modification of action variables
it turns out that meta-variables can be modified through several transitions,
possibly with distinct end states. It is thus not possible to mention that
they are unchanged in behaviors that are grouped by the state at the end of
the call to the transition function.

This commit creates specific behaviors for indicating in which case an
auxiliary variable is unchanged.