[tests] remove flamegraph test
Viewing Flamegraphs requires some specific Perl modules (e.g. open),
which are not necessarily available even with conf-perl.
Since this feature is optional and cannot be entirely tested via command-line,
removing it for now is a better approach than trying to extensively check that
each dependence is available in the test machine.