[Eva] Fixes the equality domains with option -eva-all-rounding-mode-constants.
The default comparison of Cil_datatype does not distinguish constants
with different textual representations in the source code when they
represent the same value.
However, when option -eva-all-rounding-modes-constants is enabled, the
evaluation of floating-point constants may differ according to their
textual representation: when the constant is not exactly representable in the
floating-point type, an interval is used instead of the nearest singleton
value. In this case, two constants with the same nearest value but different
textual representation may not be equal. We thus use the [compare_strict]
comparison, which also compares the textual representation of constants.