[variadic] Monomorphise the call to a variadic function even in case of errors
The previous version was using the generic translation of variadic
function if an error occurred during the translation. The problem was
that the generic translation is correct if and only if the declaration
of the function can be modified. That is not the case for variadic
functions from the libc.

This commit creates a monomorphised fallback version of the variadic
function using the types of the given parameters of the function.