[hdrck] quit when Ctrl+C is pressed
Sys.cmd is called _very_ often and thousands of times; whenever a Ctrl+C
is pressed while inside a subcommand, it returns a non-zero code. But
in many loops, a message is emitted and nothing else, so hdrck continues on
and on. Besides, the Ctrl+C leads to spurious "invalid header" messages at the
end, because of that. The fix simply makes the program terminate.