[Eva] api: Using GADT for Response.t
I've replaced the polymorphic variant with a GADT and modified what was
necessary to compile. It should have the same behaviour.

The only questionnable modification is on the type evaluation of the
Make functor. I've been forced to add a new 'callstack parameter because
Response.t has now two parameters.