[ci] remove unused .so from pathcrawler before patching elf files
It turns out that in newer nixpkgs version the (very outdated)
dependencies of these dynamic libraries do not exist anymore, leading
autoPatchelf to fail. Fortunately, since the CI server use the x86_64
versions that look for newer dependencies, we can simply remove the
offending files. This begs the question of whether these .so are
actually used in some configuration, though.