[Kernel] remove buggy pretty-printing of _Static_assert
Pretty-printing _Static_assert was already hackish for the "not-a-struct-field"
case, and for the latter, it was incorrect: conversion between Cabs expressions
and Cil expressions does not work in every case; the normalized version was
not always identical to the original one; etc.
For now, we consider _Static_assert's not to be essential in pretty-printing.
This "revert-like" commit ensures the code is available in the history, if
we need it later.