[ivette] Fixing react unmounted component bug
The Valuecomponent had lots of weird bugs, triggering the warning
<Can't perform a React state update on an unmounted component>.
This was caused by an ugly, full of unsafe effect, way to handle the
Model used by a lot of components used by the ValueComponent

Now, the Model is created only once by the ValueComponent itself, and
given as a property to every component that needs it.

As a consequence, the useModel() function is now useless.