[ivette] Bug fix in by callstack behavior
The bug was that a clic on the by callstack button would not display the
values of all callstacks, instead showing only the consolidated one and
a unique callstack. A second clic was necessary to perform the request,
update the caches and finally display the complete data.

The fix we use here is quite brutal: we invalidate the values cache in
the model. I would like to find a cleaner way.