[ivette] Visible functions filter
There was a context menu on the function folder that proposed to filter
the displayed function. It was impossible to find it without being told.
Now, there is a settings button on the right. Because of this new
button, the hide button has been replaced by a classical fold feeling,
with the arrow on the left.