[Eva] Fixes a performance issue in the Results API.
In abstract domains, adds an optional argument [selection] when getting
the states by callstack (after an analysis): only returns the states for
the given callstacks.

Fixes a performance issue in the Results API when evaluating a request in
some callstacks among many (> 1000): creating the hashtbl for all callstacks
is costly, while it is immediate for the given callstacks.