[make] avoid differences between macOS and Linux due to usage of 'echo -n'
On macOS, when running via Makefile, '-n' is not seen as an 'echo' option and
gets printed in the output. With 'echo <msg>\c', it works on macOS, but not on
Linux, which requires '-e' to interpret escape characters.
'printf' avoids the issue since it does not include a terminating newline by
default.