[Ivette] Bug fix in text field setter
The dispatched operations order is important and caused a bug that would
prevent scrolling to a marker outside of the ASTview componant if it was
in another function than the one currently displayed. The text field
setter now dispatch the document changes before the field update. This
changes impact how the MarkerScroller ViewUpdater catches an update of
the selected marker, allowing to perform the scrolling when needed.