[Kernel] support limited version of 'weak' function attribute
This attribute is used by some STM32 compilers to allow function definitions
to be overridden by other function definitions.

Mergecil does not (easily) allow removing a previous existing definition to
override with a new one (undoing the former), but we can detect it and ask the
user to change the order of sources in the command line. In practice, this is
sufficient for some case studies.