[server] Fixes updates of the property synchronized array.
The change of a property status can also impact the status of other properties
which are logical consequences of the first one (and so on: these logical
consequences can be hypotheses of yet other properties).
In the server, when a property is modified, as we can't known which other
properties are impacted, the whole synchronized array is reloaded.