[alias] Abstract_state.union: fix multiple occurrences of same variable
When forming the union of two graphs the following sequence could occur:
join a c;
join b c;
After having join a and c, c would no longer exist leading to a noop for
the second join.

Instead we need to join the set {a,b,c} at once.
This current implementation is not very efficient. It may have to be
reworked if the number of variable sets become large.