[alias] potentially problematic change to Abstract_state.call
The change makes sense. It first looks for the vertex in the function
summary and proceeds to transfer its edges to the corresponding argument
vertex only if it finds one.

However this changes some test output weirdly removing some output.
It seems that before this change, some instructions were analysed
multiple times, which is no longer the case. Everything seems to be
correct though.