[Eva] Small optimization of trace_partitioning.
Rationing of states is only applied on statement vertices. When no rationing
is performed, all states are propagated with a unique key. In this case, it is
useless to try to join duplicate keys, or to join states with previous states
from the same key.

Also, all partitioning actions that only depend on the vertex statement are
created in the [empty_store] for a vertex, and not in the [join] function.