[Eva] Octagons: fixes the [relate] function.
- Includes both direct and indirect dependencies of variables.
- Avoids using Base.SetLattice.project.
- Properly catches Not_found error when a variable is not in relation in the
  state.

Also fixes [get_var_bases] function on missing variables.
This function is called by [relate] according to the [Relations] map, which is
an over-approximation of the relations really inferred in the state — it can
contains variables that do not belong to the state, so [get_var_bases] should
not crash in that case.