[alias] add points-to interpretation to analysis
To date the API functions of this plugin only gave access to the
information wether which pointers were aliases of one another, not to
which values a given pointer may point to, even though « Steensgaard's
algorithm » on which this plugin is based is an algorithm for a
points-to analysis.
This commit adds these missing API functions and documentation.