[e-acsl] fix bug "unexpected error in \sum translation"
Crash of the program arith/extended_quantifier.c where the contract
"/*@ assert \sum(x * x, 2, \lambda integer k; k) == 0; */;" is executed.

[e-acsl] Failure: unexpected error in \sum translation.

It is a GMP problem. The type of x is unsigned long and then x*x is GMP
in a GMP context. In order to fix this bug, we have to return the GMP type
instead of an error message, in the case where get_cast returns "None".

Note: This commit reveals another bug fixed in the next commit.