[Eva] Fixes a bug on goto statement skipping local variable declarations.
All variables local to a block are now introduced in domain states as soon as
the analysis enters the block. This ensures all variables in scope have always
been introduced in domain states, even when a goto statement jumps to a block
after the declaration of some local variables, skipping their initializations
(these variables would have been introduced in the states, but not initialized).

As a downside, variables explicitly initialized at declaration enter the scope
too early, as they should be introduced on the fly when encountering their
[Local_init] instruction.