[Eva] Improves the behavior of cvalue callbacks on calls and call results.
In Cvalue_callbacks:
- Hooks registered by [register_call_hook] are now always applied at the start
  of the analysis of each function call. (It was previously applied at the end
  of the analysis when a builtin or the specification were used.)
- Hooks registered by [register_call_results_hook] are always applied at the
  end of the analysis of each function call. (It was previously not applied
  when a builtin or the specification were used.)

As a consequence, "call" hooks no longer have access to the "from" results of
builtins, but "call_results" hooks do.

The Inout and From plugins now use "call_results" instead of "call" hooks
to interpret calls when a builtin or the specification has been used by Eva.