[e-acsl] fix bug "typing was not performed on construct x"
Fixes a crash of e-acsl because of unimplemented logical functions
like sqrt in tests/idct/ieee_1180_1990.c. Running frama-c -e-acsl on
this file results in the following exception:

[e-acsl] idct/ieee_1180_1990.c:169: Warning:
E-ACSL construct `pragma' is not yet supported. Ignoring annotation.
[e-acsl] Failure: typing was not performed on construct x

The problem here is that typing of arguments of unknown logical function
is omitted.