[e-acsl] new compilation scheme via intermedia language
Introduce a new compilation scheme where E-ACSL logic terms are
translated to CIL is implemented as a two-stage process.
First, E-ACSL is translated into an intermediate language. Only then in
the second stage the intermediate language is translated into CIL.

The module Interlang specifies the E-ACSL intermediate language type,
along with pretty printing functions. As this language is still subject
to frequent changes, documentation remains scant for the moment.

The module Interlang_gen helps in the creation of Interlang expressions,
and is thus useful for the first stage. The module Interlang_trans
specifies the translation of Interlang expression to CIL and is relevant
for the second stage.
Both modules make heavy use of the RWS monad, defined in the Monad_rws
module. RWS stands for Reader, Writer, State. The RWS monad is a monad
to model computations with side-effects and environments in a purely
functional and type-safe manner.

As the new compilation scheme covers E-ACSL only partially, it fails on
many E-ACSL terms. Therefore we keep for the moment the old
direct-to-CIL compilation scheme, and, whenever the new compilation
scheme fails for a given sub-term raising the Interlang_gen.Not_covered
exception, we fall back on the direct-to-CIL compilation scheme for that
sub-term.