[Eva] Fixes message about garbled mix generated by a specification.
On assigns clause on struct, the message could be emitted even if no garbled
mix had been created, as the offsetmap for the struct was read as one cvalue,
which could be very imprecise.
Instead, this commit iters on the offsetmap and warn for each imprecise garbled
mix in the offsetmap, if any.