[alias] overhaul API
It makes no sense to reproduce all of the fold functions for sets of
variables and lvals, otherwise we should do the same for iter functions
and the like.
An API user should rather use the pre-existing functions from LSet and
VarSet. Therefore let's mark most of these functions as deprecated.

Fixes #2692