[e-acsl] statically determine some predicates for special pointers
For the special pointers stdin, stdout, stderr, and &errno the results
of \pvalid, \pinitialized, \freeable are statically known, so we return
the result directly instead of generating calls to the runtime functions.

This complements the implementation of the same semantics in the runtime
libraries in the previous commit. As the runtime implementation is
sufficient, this commit is an optimisation.

We currently do not optimise logic functions such as \block_length,
\offset, \base_addr, but we do not expect too many occurrences of terms
such as \base_addr(*(&stdin + 9)).