[e-acsl] Logic_functions: fix arity confusion of generated functions
If a function is translated multiple times with the same profile but
with different return types, then things get mixed up.
The problem is outlined in detail in the previous commit.

The solution is to take into account the return type of the generated
functions.

This commit also includes a cosmetic overhaul of the memoisation code.