[alias] test case for global variable definitions
This commit modifies one test case, such that aliases originate from
global variable definitions (outside functions). This is a pathological
case for the current state of the analysis as it does not at all take
into account such definitions, and thus misses the aliases.

This will be fixed by the subsequent commit.