[e-acsl] pathological example causing incoherent arity
This simple example causes an exception Invalid_argument("List.map2").

/*@ logic ℤ f4 (ℤ x) = x ≡ 0 ? 0 : f4(x-1) < 1<<99 ? 1 : 2; */

/*@ assert f4 (0) ≡ 0; */

The two call sites of f4 require different return behaviours of f4.
The recursive call is in a GMP context, so the result needs to be
passed back to the caller via an additional argument as a reference.
The assert calls f4 from an int context ⇒ the result is expected to be
passed back normally as an int using return.

The current code in Logic_functions maintains a hash table called
memo_tbl, which memoises for each logic function C functions that has
been generated for each encountered profile. This hash table does not
take into account the calling context (the result type and thus whether
an additional result parameter is required). Therefore, the generated
C function for f4 is erroneously reused, leading to the above exception
due to a mismatching arity.