[e-acsl] add pathological case for negative integer handling
The -rte flag causes the generation of a negative integer in a rational
context. In that case Translate_terms.constant_to_exp converts the
integer to a real string representation by converting the integer to a
string and then normalising that string using Gmp.Q.normalize_str.

However Gmp.Q.normalize_str does not correctly handle negative integers.

Note that negative integers do not occur when writing for instance -1 in
the source file as it is parsed as TUnOp(Neg, _).