[Eva] Fixes structure to allow restricting domains to functions with Mthread.
When restricting a domain to some C functions, some data is added to the
domain tree with a "void" structure. This leads to a failure when Mthread
compares domain structures to ensure it remains the same across all analyses
of all threads.

A "void" structures now requires a typed key (as any leaf): this allows
comparing two structures that include a "void" subpart.