[e-acsl] fix logic variable escaping its scope
When translating a function we must ensure that it cannot re-use any
expressions resulting from translating logic variables from its call
site. One might think that this may not occur, since logic variables are
unique. However sometimes the same logic function needs to be translated
twice (with different signatures). In that case there are indeed
indentical logic variables for different C code to be generated, and
re-using that code would be erroneous.

See the pathological example from the previous commit.

This commit fixes this by clearing the map of logic bindings before
translating a function body.