[Kernel] Fix old DoS issue on the floating point parser
The loops use to adjust the exponent, the numerator and the denominator
can take a long time (more than 200000 iterations) for numbers with a
large exponent, for instance 1e99999. This is trivially not
representable as a double but the loops are too long to figure it out.

Those loops are here because the actual fast and correct way to make
the adjustement requires to compute a number δ using binary logarithms,
and more specifically the ceiling of the difference of two binary logs.
This computation cannot be done exactly, as Zarith only provides (which
is absolutly normal, don't get me wrong) a rounded down approximation of
the binary log of any given integer. However, it can still be used to
compute an underapproximation of the exponent that will be produced by
the loops. And computing this underapproximation is trivially fast.
We can thus use this underapproximation to check if the number will
be a positive infinite in no time, avoiding the loop and thus the DoS.

An important point is that this method is correct. Proving that the
proposition "δ is larger than the maximal exponent" implies that
"the parsed number is infinity" is not hard and is underlined in the
comments of the parsing function.