[e-acsl] associate location with memoised exceptions
Up to now locations in E-ACSL are set using Current_loc.set. Since this
is stateful it is difficult to track we are replacing in the subsequent
commit the usages of Current_loc.set by a functional variant
Current_loc.with_loc or its semantic sugar let<> UpdatedCurrentLoc.

Exceptions in E-ACSL are raised in a delayed fashion:
Some modules (Typing, Interval) store raised exceptions in a memoisation
table instead of raising them immediately. Then later Error.handle will
raise the exceptions later. When tracking locations in a stateful
manner, at the moment the exception will be raised corresponds to the
latest location set, thus the latest processed language element. This is
not the case if locations are tracked in a functional manner. Therefore
we have to associate with exceptionis the location at the moment it is
originally raised (and memoised).

This commit reduces the precision of warnings of the example
tests/arith/oracle/at_on-purely-logic-variables.res.oracle.
This is due to a missing Current_loc.set/with_loc which is added in
subsequent commit.