[kernel] Add Compression and Channel library
The `Compression` library is a wrapper aroung `Gzip` with additional
helper functions.

The `Channel` library abstracts over `Stdlib` and `Compression` channels
and can be used in place of `in_channel` and `out_channel` to allow for
on-the-fly compression.

The module `Filesystem.Compressed` has been added with functions
`with_open_in_exn` and `with_open_out_exn` to be able to safely open and
close channels on compressed files.