[kernel] Fix typing of types in builtins
The builtins `__builtin_types_compatible_p` and `__builtin_va_arg` need
a type as argument. Frama-C converts them to an expression to be able to
type the builtin call.

This commit change the conversion in `cparser.mly` from a sizeof on the
type to a cast with a marker indicating from which builtin the argument
is coming from. This allows `cabs2cil` to type this expression
differently than a regular cast or sizeof.