[Kernel] Fixes inconsistent AST on inline functions (when not removed by rmtmps).
On the definition of an inline function, cabs2cil created a declaration of a
copy of the function, which was not properly registered and led to a failure
of the AST integrity check:
"function <name> is supposed to be defined, but not registered as such."

The copied declaration was unused and removed by Rmtmps in most cases,
thus preventing the failure, unless the inlined function had a specification
or "-keep-unused-functions all" was set.

This commit removes this copied declaration of inline functions.