[Kernel] Better explanation in filters invariant computation
In the invariant computation, two approaches are used to compute limits
of infinite geometric series in a linear space. The first one is the
exact approach that does only relies on linear computation. The second
requires to compute matrices norms, which is ugly but necessary in that
particular case. The explanation of why wasn't clear enough and is now
improved.