[Eva] A generic approach to the IEEE-754 semantics
Numerors implements an abstraction of the IEEE-754 semantic using
intervals. However, nothing per se restrict us to only use intervals.
In fact, any abstraction, as long as it provides some arithmetic
operators and dedicated tools, can be used to build an abstraction
of the IEEE-754 semantic.

This huge commit is the result of hours of development in different
branches, here condensed into a final, clean form, which is why
everything kind of pop out of nowhere. I know it is not easy to
treat such a huge chunck of work, but it seems to me that this
is the most reasonnable compromise. I will try to give an overview
of the files added by this commit.

The main point is the {!IEEE754} module (provided with a ml and a mli)
that provides a functor that build an abstract value based on some
abstract "modeling", i.e a set of abstractions and tools used to
represent the IEEE-754 semantic and compute the reduced product between
absolute and relative error bounds. Here, I mean abstract value in the
sens of Eva abstract values, i.e a module that implements the
{!Abstract_value.Leaf} interface. The returned module also provides
different arithmetic operators and utilitary functions over the
abstract values.

The {!IEEE754_sig} interface (provided as a ml only module) exports all
the types and signatures that are used to build the abstract values.
They are separated to avoid dupplications.